+++ bindings/go/evmc/evmc.go
@@ -15,9 +15,9 @@ package evmc
 #include <stdlib.h>
 #include <string.h>
 
-static inline enum evmc_set_option_result set_option(struct evmc_instance* instance, char* name, char* value)
+static inline enum evmc_set_option_result set_option(struct evmc_vm* vm, char* name, char* value)
 {
-	enum evmc_set_option_result ret = evmc_set_option(instance, name, value);
+	enum evmc_set_option_result ret = evmc_set_option(vm, name, value);
 	free(name);
 	free(value);
 	return ret;
@@ -31,7 +31,7 @@ struct extended_context
 
 extern const struct evmc_host_interface evmc_go_host;
 
-static struct evmc_result execute_wrapper(struct evmc_instance* instance,
+static struct evmc_result execute_wrapper(struct evmc_vm* vm,
 	int64_t context_index, enum evmc_revision rev,
 	enum evmc_call_kind kind, uint32_t flags, int32_t depth, int64_t gas,
 	const evmc_address* destination, const evmc_address* sender,
@@ -52,7 +52,7 @@ static struct evmc_result execute_wrapper(struct evmc_instance* instance,
 	};
 
 	struct extended_context ctx = {{&evmc_go_host}, context_index};
-	return evmc_execute(instance, &ctx.context, rev, &msg, code, code_size);
+	return evmc_execute(vm, &ctx.context, rev, &msg, code, code_size);
 }
 */
 import "C"
@@ -144,18 +144,18 @@ const (
 	Istanbul         Revision = C.EVMC_ISTANBUL
 )
 
-type Instance struct {
-	handle *C.struct_evmc_instance
+type VM struct {
+	handle *C.struct_evmc_vm
 }
 
-func Load(filename string) (instance *Instance, err error) {
+func Load(filename string) (vm *VM, err error) {
 	cfilename := C.CString(filename)
 	var loaderErr C.enum_evmc_loader_error_code
 	handle := C.evmc_load_and_create(cfilename, &loaderErr)
 	C.free(unsafe.Pointer(cfilename))
 
 	if loaderErr == C.EVMC_LOADER_SUCCESS {
-		instance = &Instance{handle}
+		vm = &VM{handle}
 	} else {
 		errMsg := C.evmc_last_error_msg()
 		if errMsg != nil {
@@ -165,17 +165,17 @@ func Load(filename string) (instance *Instance, err error) {
 		}
 	}
 
-	return instance, err
+	return vm, err
 }
 
-func LoadAndConfigure(config string) (instance *Instance, err error) {
+func LoadAndConfigure(config string) (vm *VM, err error) {
 	cconfig := C.CString(config)
 	var loaderErr C.enum_evmc_loader_error_code
 	handle := C.evmc_load_and_configure(cconfig, &loaderErr)
 	C.free(unsafe.Pointer(cconfig))
 
 	if loaderErr == C.EVMC_LOADER_SUCCESS {
-		instance = &Instance{handle}
+		vm = &VM{handle}
 	} else {
 		errMsg := C.evmc_last_error_msg()
 		if errMsg != nil {
@@ -185,21 +185,21 @@ func LoadAndConfigure(config string) (instance *Instance, err error) {
 		}
 	}
 
-	return instance, err
+	return vm, err
 }
 
-func (instance *Instance) Destroy() {
-	C.evmc_destroy(instance.handle)
+func (vm *VM) Destroy() {
+	C.evmc_destroy(vm.handle)
 }
 
-func (instance *Instance) Name() string {
-	// TODO: consider using C.evmc_vm_name(instance.handle)
-	return C.GoString(instance.handle.name)
+func (vm *VM) Name() string {
+	// TODO: consider using C.evmc_vm_name(vm.handle)
+	return C.GoString(vm.handle.name)
 }
 
-func (instance *Instance) Version() string {
-	// TODO: consider using C.evmc_vm_version(instance.handle)
-	return C.GoString(instance.handle.version)
+func (vm *VM) Version() string {
+	// TODO: consider using C.evmc_vm_version(vm.handle)
+	return C.GoString(vm.handle.version)
 }
 
 type Capability uint32
@@ -209,13 +209,13 @@ const (
 	CapabilityEWASM Capability = C.EVMC_CAPABILITY_EWASM
 )
 
-func (instance *Instance) HasCapability(capability Capability) bool {
-	return bool(C.evmc_vm_has_capability(instance.handle, uint32(capability)))
+func (vm *VM) HasCapability(capability Capability) bool {
+	return bool(C.evmc_vm_has_capability(vm.handle, uint32(capability)))
 }
 
-func (instance *Instance) SetOption(name string, value string) (err error) {
+func (vm *VM) SetOption(name string, value string) (err error) {
 
-	r := C.set_option(instance.handle, C.CString(name), C.CString(value))
+	r := C.set_option(vm.handle, C.CString(name), C.CString(value))
 	switch r {
 	case C.EVMC_SET_OPTION_INVALID_NAME:
 		err = fmt.Errorf("evmc: option '%s' not accepted", name)
@@ -226,7 +226,7 @@ func (instance *Instance) SetOption(name string, value string) (err error) {
 	return err
 }
 
-func (instance *Instance) Execute(ctx HostContext, rev Revision,
+func (vm *VM) Execute(ctx HostContext, rev Revision,
 	kind CallKind, static bool, depth int, gas int64,
 	destination common.Address, sender common.Address, input []byte, value common.Hash,
 	code []byte, create2Salt common.Hash) (output []byte, gasLeft int64, err error) {
@@ -242,7 +242,7 @@ func (instance *Instance) Execute(ctx HostContext, rev Revision,
 	evmcSender := evmcAddress(sender)
 	evmcValue := evmcBytes32(value)
 	evmcCreate2Salt := evmcBytes32(create2Salt)
-	result := C.execute_wrapper(instance.handle, C.int64_t(ctxId), uint32(rev),
+	result := C.execute_wrapper(vm.handle, C.int64_t(ctxId), uint32(rev),
 		C.enum_evmc_call_kind(kind), flags, C.int32_t(depth), C.int64_t(gas),
 		&evmcDestination, &evmcSender, bytesPtr(input), C.size_t(len(input)), &evmcValue,
 		bytesPtr(code), C.size_t(len(code)), &evmcCreate2Salt)