+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning].
   
 ### Changed
 
+- The `evmc_instance` renamed to `evmc_vm`.
+  [[#430](https://github.com/ethereum/evmc/pull/430)]
 - The `evmc::vm` renamed to `evmc::VM` in C++ API.
   [[#252](https://github.com/ethereum/evmc/pull/252)]
 - Previously deprecated `helpers.hpp` header file has been removed.